Add support for preserving product images with WC1C_PRESERVE_PRODUCT_IMAGES constant
git-svn-id: https://plugins.svn.wordpress.org/woocommerce-and-1centerprise-data-exchange/trunk@1110989 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ exchange/import.php
@@ -460,6 +460,8 @@ function wc1c_replace_post_attachments($post_id, $attachments) {
       }
     }
 
+    if (defined('WC1C_PRESERVE_PRODUCT_IMAGES') && WC1C_PRESERVE_PRODUCT_IMAGES) continue;
+
     $result = wp_delete_attachment($post_attachment->ID);
     if ($result === false) wc1c_error("Failed to delete post attachment");
   }
